bot: Propose a new method for bot access controls
Instead of relying on PR labels to indicate whether bot functionality
should be allowed, use a commit status entry instead. This allows us
to be confident that access will get reset when changes are made to
the PR, which should require another check before the bot can be run
again.

Signed-off-by: Russell Bryant <rbryant@redhat.com>