DRAFT: Capture thoughts around a future backend service
We now have a functional backend pipeline operated by our backend
maintainers team. As we look to the future, we should try to
articulate the key attributes we envision in a future we aspire to.

The current backend works for the InstructLab project. While this
backend works for us, we want a flexible backend that works for many.

This document is only a starting point and does not represent a
complete scoping for the future. The most important section proposed
here is the Guiding Principles section. We should identify the set of
key priciplles that guide us as a project as we design for the future.

Signed-off-by: Russell Bryant <rbryant@redhat.com>