Propose a backend working group
There is a broad group of people interested in collaborating on the
software needed to scale out the workflow in a clustered environment.
As a first step toward formalizing this work, this document describes
a working group that will collaborate on documents in this repository
to defined high level architecture and design for the related
components to build out this updated backend infrastructure.

Signed-off-by: Russell Bryant <rbryant@redhat.com>