Create `docs/cli` directory
This PR creates a directory for CLI-focused design documents to help
keep our design repo a bit more organized.

I also updated the README to clarify that in practice we're using this
as a repo of design artifacts for all of InstructLab and not just
changes that cross multiple repositories. This can always evolve over
time, but it reflects our current reality right now.

Signed-off-by: Russell Bryant <rbryant@redhat.com>