amend merge criteria
the rule that the Oversight Committee needs to merge every dev-doc PR has not scaled with the project. We now write dev-docs for most design changes for any instructlab repo. This rule has also not regularly been followed.
Amend this rule such that 2 maintainers of any instructlab can approve a PR and a third maintainer of any instructlab repo can verify that sufficient reviews have been given and merge a PR

Signed-off-by: Charlie Doern <cdoern@redhat.com>