amend merge criteria
the rule that the Oversight Committee needs to merge every dev-doc PR has not scaled with the project. We now write dev-docs for most design changes for any instructlab repo. This rule has also not regularly been followed.
Amend this rule such that and 2 maintainers of any instructlab repo can approve and merge, and an OC member can be brought in as well if needed.

Signed-off-by: Charlie Doern <cdoern@redhat.com>