Remove requirement to use harden-runner action
In all the time it was applied, no single repo cared enough to enable
`block` policy for egress network traffic. All the action did is it
produced churn where we were updating tags for the action to new
versions.

We *could* switch these actions to `block` but I'd rather remove the
requirement if in all this time we haven't seen an actual need.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>